api/pyproject.toml
[project]
name = "studyforge-api"
version = "0.1.0"
description = "AI-powered study assistant API built with FastAPI, PostgreSQL, JWT, and LLM integrations."
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.2.0",
    "psycopg[binary]>=3.2.0",
    "pwdlib[argon2]>=0.2.1",
    "openai>=1.0.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
]
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "anyio>=4.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0"
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = [
    "-p",
    "no:cacheprovider",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*"]
